refactor modulo check into separate class
