+++ autograder-core/src/main/java/de/firemage/autograder/core/ProblemType.java
@@ -102,7 +102,7 @@ public enum ProblemType {
     COMMON_REIMPLEMENTATION_ADD_ALL,
     COMMON_REIMPLEMENTATION_ADD_ENUM_VALUES,
     COMMON_REIMPLEMENTATION_ARRAYS_FILL,
-    COMMON_REIMPLEMENTATION_MODULO,
+    USE_MODULO_OPERATOR,
     COMMON_REIMPLEMENTATION_SUBLIST,
     COMMON_REIMPLEMENTATION_ITERABLE_DUPLICATES,
     ABSTRACT_CLASS_WITHOUT_ABSTRACT_METHOD,
