+++ sample_config.yaml
@@ -129,7 +129,6 @@
 - COMMON_REIMPLEMENTATION_MAX_MIN
 - COMMON_REIMPLEMENTATION_ADD_ALL
 - COMMON_REIMPLEMENTATION_ARRAYS_FILL
-- COMMON_REIMPLEMENTATION_MODULO
 - EXCEPTION_SHOULD_NEVER_BE_CAUGHT
 - USE_DIFFERENT_VISIBILITY_PUBLIC_FIELD
 - MULTI_THREADING
@@ -155,3 +154,4 @@
 - COLLECTIONS_N_COPIES
 - REDUNDANT_ELSE
 - COLLECTION_ADD_ALL
+- USE_MODULO_OPERATOR
